LPD-6742 Renaming module and adapting it to read the exported files on the other Virtual Instance Migrator tool
