+++ modules/util/portal-tools-db-partition-virtual-instance-migrator-validator/scripts/db_partition_virtual_instance_migrator_validator.bat
@@ -1,0 +1,11 @@
+@echo off
+
+pushd "%~dp0"
+
+path %PATH%;%JAVA_HOME%\bin
+
+java -jar com.liferay.portal.tools.db.partition.virtual.instance.migrator.validator.jar %*
+
+popd
+
+@echo on
