+++ modules/util/portal-tools-db-partition-virtual-instance-migrator-validator/source-formatter-suppressions.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0"?>
+
+<suppressions>
+	<checkstyle>
+		<suppress checks="CompanyIterationCheck" files="src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migrator/validator/util/Validator\.java" />
+	</checkstyle>
+</suppressions>
