+++ modules/util/portal-tools-db-partition-virtual-instance-migrator-validator/src/testFunctional/macros/DBPartitionMigratorTool.macro
