+++ modules/util/portal-tools-db-partition-virtual-instance-migrator-validator/test.properties
@@ -1,0 +1,38 @@
+##
+## Modules
+##
+
+    modules.includes.required.test.batch.class.names.includes=\
+        **/portal-tools-db-partition-virtual-instance-migrator-validator/**/*Test.java
+
+##
+## Test Batch
+##
+
+    #
+    # Portal Hotfix Release
+    #
+
+    test.batch.run.property.query[functional-bundle-tomcat-mysql57-jdk8][test-portal-hotfix-release]=\
+        ${test.batch.run.property.query[functional-tomcat*-mysql*-jdk8][relevant]}
+
+    #
+    # Relevant
+    #
+
+    test.batch.run.property.query[functional-tomcat*-mysql*-jdk8][relevant]=\
+        (portal.acceptance == true) AND \
+        (\
+            (testray.component.names ~ "Database Partitioning") OR \
+            (testray.main.component.name ~ "Database Partitioning")\
+        )
+
+    test.batch.run.property.query[functional-upgrade*-tomcat*-mysql*-jdk8][relevant]=\
+        (portal.acceptance == true) AND \
+        (testray.main.component.name ~ "Database Partitioning")
+
+##
+## Testray
+##
+
+    testray.main.component.name=Database Partitioning
