--- modules/util/portal-tools-db-partition-virtual-instance-migrator/build.gradle
@@ -1,43 +1,0 @@
-dependencies {
-	compileInclude group: "com.mysql", name: "mysql-connector-j", version: "8.3.0"
-	compileInclude group: "commons-cli", name: "commons-cli", version: "1.4"
-
-	compileOnly group: "com.liferay", name: "com.liferay.petra.function", version: "5.5.0"
-	compileOnly group: "com.liferay", name: "com.liferay.petra.lang", version: "1.0.0"
-	compileOnly group: "com.liferay", name: "com.liferay.petra.string", version: "5.3.0"
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "org.mockito", name: "mockito-core", version: "4.5.1"
-}
-
-deploy {
-	from "scripts"
-}
-
-distZip {
-	filesMatching("**/*.jar") {
-		it.path = it.name
-	}
-}
-
-distributions {
-	main {
-		contents {
-			from "scripts"
-			into "/"
-		}
-	}
-}
-
-jar {
-	archiveName "$baseName" + ".jar"
-}
-
-liferayOSGi {
-	expandCompileInclude = true
-}
-
-startScripts {
-	onlyIf {
-		false
-	}
-}
