--- modules/util/portal-tools-db-partition-virtual-instance-migrator/source-formatter-suppressions.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0"?>
-
-<suppressions>
-	<checkstyle>
-		<suppress checks="CompanyIterationCheck" files="src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migrator/util/DatabaseUtil\.java" />
-	</checkstyle>
-</suppressions>
